use react-editext for inline editing - styling and fixes to come
