+++ package.json
@@ -257,6 +257,7 @@
     "react": "^18.1.0",
     "react-bootstrap": "^2.0.2",
     "react-dom": "^18.1.0",
+    "react-editext": "^4.2.1",
     "react-outside-click-handler": "^1.3.0",
     "react-redux": "^8.0.1",
     "react-router": "^6.2.2",
